+++ include/test/animation.h
@@ -32,7 +32,7 @@ class AnimationTest : public games::Game {
     bg->draw(display);
     block->draw(display);
     flicker->draw(display);
-    for (int i = 0; i < display->numStrips; i++) {
+    for (size_t i = 0; i < display->numStrips; i++) {
       display->strips[i][BoundaryBase].setHSV(125, 255, 255);
       display->strips[i][BoundaryBase + BoundaryHeight].setHSV(125, 255, 255);
     }
